# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neuraal API",
    "version": "0.1.0",
    "description": "Backend API for Neuraal — task/note management with AI features (summaries, auto-topic classification)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "DevUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Temporary dev-only auth. Sends a fixed user ID. Will be replaced by BearerAuth (JWT) in production."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Future JWT auth (not enforced yet)."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Resource UUID"
      }
    }
  },
  "security": [
    {
      "DevUserIdHeader": []
    }
  ],
  "paths": {
...
    "/api/topics/{id}/embedding/rebuild": {
      "post": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Rebuild topic embedding",
        "operationId": "rebuildTopicEmbedding",
        "description": "Recalculates the embedding vector for a topic using Ollama. Useful after renaming.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Embedding rebuilt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "topicId",
                    "embeddingUpdatedAt"
                  ],
                  "properties": {
                    "topicId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "embeddingUpdatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}